Move heartbeat recovery side effects into service
